+++ hapi-fhir-jpaserver-base/src/main/java/ca/uhn/fhir/jpa/dao/index/SearchParamWithInlineReferencesExtractor.java
@@ -153,117 +153,11 @@ public class SearchParamWithInlineReferencesExtractor {
 	}
 
 	private void extractCompositeStringUniques(ResourceTable theEntity, ResourceIndexedSearchParams theParams) {
+		Set<ResourceIndexedComboStringUnique> comboUniques = mySearchParamExtractorService.extractSearchParamComboUnique(theEntity, theParams);
+		theParams.myComboStringUniques.addAll(comboUniques);
 
-		final String resourceType = theEntity.getResourceType();
-		List<RuntimeSearchParam> comboSearchParams = mySearchParamRegistry.getActiveComboSearchParams(resourceType);
-
-		for (RuntimeSearchParam next : comboSearchParams) {
-			switch (Objects.requireNonNull(next.getComboSearchParamType())) {
-				case UNIQUE:
-					extractComboUniqueParam(theEntity, theParams, resourceType, next);
-					break;
-				case NON_UNIQUE:
-					extractComboNonUniqueParam(theEntity, theParams, resourceType, next);
-			}
-		}
-	}
-
-	private void extractComboNonUniqueParam(ResourceTable theEntity, ResourceIndexedSearchParams theParams, String theResourceType, RuntimeSearchParam theParam) {
-		Set<String> queryStringsToPopulate = extractParameterCombinationsForComboParam(theParams, theResourceType, theParam);
-
-		for (String nextQueryString : queryStringsToPopulate) {
-			ourLog.trace("Adding composite unique SP: {}", nextQueryString);
-			theParams.myComboTokenNonUnique.add(new ResourceIndexedComboTokenNonUnique(myPartitionSettings, theEntity, nextQueryString));
-		}
-	}
-
-	private void extractComboUniqueParam(ResourceTable theEntity, ResourceIndexedSearchParams theParams, String theResourceType, RuntimeSearchParam theParam) {
-		Set<String> queryStringsToPopulate = extractParameterCombinationsForComboParam(theParams, theResourceType, theParam);
-
-		for (String nextQueryString : queryStringsToPopulate) {
-			ourLog.trace("Adding composite unique SP: {}", nextQueryString);
-			theParams.myComboStringUniques.add(new ResourceIndexedComboStringUnique(theEntity, nextQueryString, theParam.getId()));
-		}
-	}
-
-	@Nonnull
-	private Set<String> extractParameterCombinationsForComboParam(ResourceIndexedSearchParams theParams, String theResourceType, RuntimeSearchParam theParam) {
-		List<List<String>> partsChoices = new ArrayList<>();
-
-		List<RuntimeSearchParam> compositeComponents = JpaParamUtil.resolveComponentParameters(mySearchParamRegistry, theParam);
-		for (RuntimeSearchParam nextCompositeOf : compositeComponents) {
-			Collection<? extends BaseResourceIndexedSearchParam> paramsListForCompositePart = findParameterIndexes(theParams, nextCompositeOf);
-
-			Collection<ResourceLink> linksForCompositePart = null;
-			switch (nextCompositeOf.getParamType()) {
-				case REFERENCE:
-					linksForCompositePart = theParams.myLinks;
-					break;
-				case NUMBER:
-				case DATE:
-				case STRING:
-				case TOKEN:
-				case QUANTITY:
-				case URI:
-				case SPECIAL:
-				case COMPOSITE:
-				case HAS:
-					break;
-			}
-
-			Collection<String> linksForCompositePartWantPaths = null;
-			switch (nextCompositeOf.getParamType()) {
-				case REFERENCE:
-					linksForCompositePartWantPaths = new HashSet<>(nextCompositeOf.getPathsSplit());
-					break;
-				case NUMBER:
-				case DATE:
-				case STRING:
-				case TOKEN:
-				case QUANTITY:
-				case URI:
-				case SPECIAL:
-				case COMPOSITE:
-				case HAS:
-					break;
-			}
-
-			ArrayList<String> nextChoicesList = new ArrayList<>();
-			partsChoices.add(nextChoicesList);
-
-			String key = UrlUtil.escapeUrlParam(nextCompositeOf.getName());
-			if (paramsListForCompositePart != null) {
-				for (BaseResourceIndexedSearchParam nextParam : paramsListForCompositePart) {
-					IQueryParameterType nextParamAsClientParam = nextParam.toQueryParameterType();
-					String value = nextParamAsClientParam.getValueAsQueryToken(myContext);
-
-					RuntimeSearchParam param = mySearchParamRegistry.getActiveSearchParam(theResourceType, key);
-					if (theParam.getComboSearchParamType() == ComboSearchParamType.NON_UNIQUE && param != null && param.getParamType() == RestSearchParameterTypeEnum.STRING) {
-						value = StringUtil.normalizeStringForSearchIndexing(value);
-					}
-
-					if (isNotBlank(value)) {
-						value = UrlUtil.escapeUrlParam(value);
-						nextChoicesList.add(key + "=" + value);
-					}
-				}
-			}
-			if (linksForCompositePart != null) {
-				for (ResourceLink nextLink : linksForCompositePart) {
-					if (linksForCompositePartWantPaths.contains(nextLink.getSourcePath())) {
-						assert isNotBlank(nextLink.getTargetResourceType());
-						assert isNotBlank(nextLink.getTargetResourceId());
-						String value = nextLink.getTargetResourceType() + "/" + nextLink.getTargetResourceId();
-						if (isNotBlank(value)) {
-							value = UrlUtil.escapeUrlParam(value);
-							nextChoicesList.add(key + "=" + value);
-						}
-					}
-				}
-			}
-		}
-
-		return ResourceIndexedSearchParams.extractCompositeStringUniquesValueChains(theResourceType, partsChoices);
+		Set<ResourceIndexedComboTokenNonUnique> comboNonUniques = mySearchParamExtractorService.extractSearchParamComboNonUnique(theEntity, theParams);
+		theParams.myComboTokenNonUnique.addAll(comboNonUniques);
 	}
 
 	@Nullable
